More info about updating metadata & elbfi -> elbf
Why we previously called EncryptedLocalBoxFile
as elbfi, while EncryptedRemoteBoxFile was always
the erbf? Sounds like a good question! This is only
because of *LocalBoxFolder, that was before the
presenting *LocalBoxDirectory class, so the folder
was "elbf" and the file -- "elbfi", not to confuse.

Now all elbfi should be renamed to elbf. Some very
important information here, ya? XD