TelegramAccount -> (more useful) TelegramClient
This commit says bye-bye to the old and useless
TelegramAccount class. We replace it with the
tgbox.api.TelegramClient, which is inheritor
of the telethon.TelegramClient, so it is
support all the features that client of
Telethon does, adding a little bit of
own methods, like log_in, tgboxes,
send_code, resend_code.

Now you can access your account object with
the "tc" (TelegramClient) property. Before
it was the "ta" (TelegramAccount).

Now you can even use the telethon.TelegramClient
with TGBOX! Just make sure you're already signed
in your account, because TGBOX Client uses a
slightly easier login routine (log_in method
instead of the telethon's sign_in method).

This may break your code for a little bit,
sorry for this :) Just replace a couple
of old methods names with new and ta->tc.