TGBOX update 22/22
Title doesn't make big sense, ye?

This update includes:
 1. Fingerprints: now you can't upload two files on
    the same file path with the same name. Otherwise
    is strange as far i see;

 2. Flexible LocalBox: starting from now database
    structure will auto update on a first commit
    according to the version requirements;

 3. Custom defaults: now TGBOX will fetch default
    values from the new DEFAULTS table of the
    LocalBox sqlite database, so starting from
    now you can easily change i.e METADATA_MAX
    to the more than 1MB of size. To sync your
    personal defaults with the RemoteBox you're
    advised to decrypt an EncryptedRemoteBox
    object with the DecryptedLocalBox object,
    and not with keys (BaseKey/MainKey), cuz
    otherwise RemoteBox will use the default
    defaults from the defaults module (yeah!);

 4. Proxy: we added cute and easy way to use
    proxy for any TelegramClient connections;

 5. exported -> imported: prior this update
    files that was imported from the other
    RemoteBox was called... exported? IDK
    what this is all about was, so now you
    can access this info via imported flag;

 6. Some other little improvements & bugfixes.

This update took really long because of my
life circumstances and the new feature i
tried to implement called "Mirroring". The
Mirroring was about making links to already
uploaded files. I.e if you uploaded before
a file A on a path X and try to upload the
same file A but on a path Y then we don't
upload a file again, but make a link to
it (mirror) and push to RemoteBox. Sounds
pretty good? Well, i think so. But there
is really a big amount problems behind
this: i.e how to clone it, how to remove
it, how to share it, etc, so i abandoned
this idea. Telegram isn't really designed
for such things in my humble opinion.