Init ELBF and DLBF with the ELB & DLB
Prior to this commit the EncryptedLocalBoxFile was
initialized by the TgboxDB from which we was making
the *LocalBox. That's a HUGE and stupid overhead.

Now the *LocalBoxFile classes is more consistent
and, what's more important, iteration speed over
LocalBox should increase thanks to this updates.

There is also a couple of little updates to the
other things, like more log-strings, etc...