Change version to v1.2-indev
We should immediately change version in indev branch
after release (with first commit to indev) to
distinguish indev from release. Now this is rule :)

So, right before the release of v1.2 we will remove
the "-indev" suffix, create a tag & release, then
make a v1.3-indev (if it's expected as big update,
otherwise we can create a v1.2.1-indev version)

That's it! Maybe...