Make TelegramVirtualFile the class-wrapper
TVF allow to upload file from the Telegram chat (that already
uploaded). Prior to this commit, we create separate Telegram
connection for the every TVF object. This is stupid.

Now you can import TelegramVirtualFile from the
tgbox.api.utils module, wrap Document in it,
attach the TelegramClient (i.e from drb)
and use it on the dlb.prepare_file(...).