Fix Python ECDSA RequestKey/ShareKey & CHANGE THEM
(0)

Bugs with Pure-Python ECDSA (when 'cryptography' lib
is not presented we use 'ecdsa' library) after
version bump was fixed.

(1)

Prior to this commit we used resulted ECDH shared secret
to encrypt key-in-transfer as is. Now we firstly hash
the shared secret bytes with the SHA256. This is
not critical (especially in TGBOX), but recommended,
so i implemented it.

(2)

Also, before to this commit the Giver made the private
key from the sha256(mainkey + {file||box}_salt). Here's
the problem: for any new same Box||File Requester the
part of ShareKey (the public key of Giver) was
identical. This doesn't expose ANY security risks, but
Requester1 with which you already shared some file
may know that you share the same file with
Requester2, and nothing more.

To fix this (non security, rather {potentially)
privacy) flaw we will now make a Giver private
key with the Requester's RequestKey as follows:

> sha256(mainkey + sha256({file||box}_salt + requestkey))

This will guarantee that any unique Requester will
receive a totally unique (for him) ShareKey, because
RequestKey that Requester makes depends on MainKey.

(#)

This is NOT security-related thing. I just wanted
to improve this flaws. This commit will break
backward compatibility for RequestKey/ShareKey
with TGBOX versions < 1.2. This is minor, so
major version (1) will not be incremented.