DirectoryKey, updated Key making funcs, etc
This commit introduce the new Key: the DirectoryKey!

Prior to this commit we was making the file encryption
keys (FileKey) with the MainKey plus FileSalt. Because
of this, to share a many files separately with the
other people we needed to make a ShareKey for *each*
of the target Box files. This can be a long walk.

Started from this commit, we will make the DirectoryKey
from the head PartID of the *Box* file (see docs for
:How does we store file paths: chapter) and MainKey
as follows: SHA256(SHA256(MainKey) + PartID) =
= DirectoryKey (see tgbox.keys.make_dirkey func).

We will encrypt the resulted DirectoryKey with
MainKey and will place it in a public part of
the Metadata as `edirkey`.

With the DirectoryKey and FileSalt we will make
a FileKey. This change will allow us to share
the DirectoryKey (as ShareKey) with some other
user and give it access to the *whole* directory
of the *Box* files. In fact, the DirectoryKey
will *NOT* encrypt anything, it will be used
only to derive filekeys.

The DirectoryKey will be later explained in
docs, stay tuned if you're interested.

v1.3+ will support legacy filekeys, but <1.3 will
not, so make sure to use only 1.3+ after you're
uploaded some files from this version.

This is, again, some minor breaking change. Sorry
for these, i hope it will be last :)

... and some other changes here:
    * Rewrite & update the key making functions
    * Unpack Metadata only in Encrypted* class
    * Better support for decrypt and erase_encrypted_metadata
        kwargs in .get_file(...), .files(...)

    * get_sharekey in DecryptedLocalBoxDirectory
    * some updates to docstrings
    * etc...