Reverse in *LocalBox.files() & more
Now *LocalBox.files() generator support reverse
flag so you can iterate from the newest files
to oldest.

The default behaviour for *RemoteBox.files()
generator was changed. Previously it yields
a files from the newest to oldest, and was
opposite to the *LocalBox.files(), which is
the oldest to newest by default. Now, the
*RemoteBox.files() and *LocalBox.files()
both by default yield Oldest -> Newest
and support 'reverse' boolean flag.