is_encrypted, revisited box_path, box_name see ...
Now Encrypted* and Decrypted* classes have 'is_encrypted'
property. Return from Decrypted* is always `False`.

I made a little revisit of the `make_localbox` and
the `clone_remotebox` functions. Now we can specify
the 'box_path' and 'box_name' arguments separately
and in obvious way. Previously, clone_remotebox was
expecting 'box_path' to also have a filename in it,
now it should be specified in 'box_name'. This is
the only _minor_ breaking change in this commit.

The `clone_remotebox` now is a more smart. It will
firstly check if it's possible to decrypt at least
one file, and if not, raise AESError. The previous
behaviour is iterate over whole RemoteBox and import
nothing. Such a waste!

Started from this commit, if you will provide
the original Phrase->BaseKey on RemoteBox clone,
your LocalBox will not store encrypted MainKey,
because we can make it from BaseKey.

(Currently i forgot to well test the last change,
so you may expect some commits with fixes on it xD)