Return the Sphinx RTD Theme
The Sphinx Book Theme presented in the v1.3 docs update
is cute but just AWFUL in code functions search, which,
in my opinion, is more important over look. I'd like to
update the Docs theme, but currently can not find the
any other that i will like, so let's return the old
well-known one.