Protection against the Bit-flipping attack (HMAC)
This experimental changes enable the protection against the
Bit-flipping attack. In short: BFA is an attack on CBC where
attacker can change the parts of ciphertext in a controlled
way to inject desired bytes. Details will be explained in the
v1.5 docs (if this branch will be merged with indev). As for
now, see Wiki: en.wikipedia.org/wiki/Bit-flipping_attack

::SHORT NOTE::

While this attack is pretty serious on a network protocols,
here in TGBOX it's a very little breach that is *almost*
impossible to use, as it require write(/edit) access to
your encrypted files in Remote Telegram Channel. It can
be made either by folks who control Telegram servers or
by someone with enough Channel-Admin rights. However,
if we *can* fix weaknesses (even little), -- we *should*.

ADDITIONS:

Now on file upload we will compute HMAC-SHA256 (over plaintext!)
and place it at the end of encrypted file, as:
    <fixed_metadata><metadata><encrypted_file><hmac>

For such addition TGBOX Protocol now offers the
'tgbox.keys.make_hmackey' and the respective
'tgbox.keys.HMACKey' class, which will be used
to compute a HMAC of plaintext bytes.

On file download we will compute HMAC AND compare
it with HMAC attached to EncryptedRemoteBoxFile. If
it's not the same, -- error will be raised.

[tgbox/keys.py]
- The new make_hmackey() func & HMACKey() class

[tgbox/api/utils.py]
- PreparedFile() now require 'hmackey'

[tgbox/tools.py]
- Now PackedAttributes() key/value order is random by default
- OpenPretender() now require 'hmac_state' ('hmac.HMAC')
- OpenPretender() will auto-compute and attach HMAC now
- OpenPretender() now will always return b'' after full
  target file encryption instead of useless bytes
- Now prbg() is not linked to global 'random' state

[tgbox/api/local.py & tgbox/api/remote.py]
- Now first block of Secret Metadata (IV) is Protected with
  the new '_BFP' field (5 random bytes, 16 in total due to
  way PackedAttributes algo works). Any other Bit-flip in
  PackedAttributes will corrupt it and result in error.
- Updated Metadata now will also have a '_BFP' field at
  the start to protect IV. '_BFP' is internal and will
  be NOT presented in the '*.residual_metadata'
- Now Secret Metadata has new 'has_hmac_sha256' field,
  which indicate that we should compute the HMAC for
  file. Do Not otherwise for backward compatibility.
- New 'has_hmac_sha256' property on *BoxFile
- Now 'DecryptedRemoteBoxFile.download()' will compute
  and compare HMAC if 'has_hmac_sha256' presented, but
  this behaviour can be omited with 'omit_hmac_check'.

This is a QUICK INSIGHT into this commit. Details will
be described in Documentation of v1.5 in future.