Code cleanup plus some fixes
Mostly this commit consits of changing the
''' to ' in strings. As I already said in
CLI, I don't know why I used them for so
long while we can just type ' instead.

Let's hope that this changes will not break
anything important. Shall we?

There is a pair of small bugfixes too.